+++ datalab/tests/features/utilities/logview_error.py
@@ -9,14 +9,14 @@ Log viewer test: raise an exception and create a seg fault in DataLab
 from guidata.qthelpers import qt_app_context
 
 from datalab.env import execenv
-from datalab.gui.main import CDLMainWindow
+from datalab.gui.main import DLMainWindow
 
 
 def error():
     """Raise an exception and create a seg fault in DataLab"""
     with execenv.context(unattended=True):
         with qt_app_context(exec_loop=True):
-            win = CDLMainWindow()
+            win = DLMainWindow()
             win.test_segfault_error()